refector: 요청 관련 type을 shared계층으로 이동
